% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logging.R
\name{addDefaultConsoleLogger}
\alias{addDefaultConsoleLogger}
\title{Add the default console logger}
\usage{
addDefaultConsoleLogger()
}
\description{
Add the default console logger
}
\details{
Creates a logger that writes to the console using the "INFO" threshold and the
\code{\link{layoutSimple}} layout.
}
\examples{
logger <- addDefaultConsoleLogger()
registerLogger(logger)                      
logTrace("This event is below the threshold (INFO)")
logInfo("Hello world")                       
unregisterLogger(logger)  

}
