% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loggers.R
\name{addDefaultErrorReportLogger}
\alias{addDefaultErrorReportLogger}
\title{Add the default error report logger}
\usage{
addDefaultErrorReportLogger(fileName = file.path(getwd(), "errorReportR.txt"))
}
\arguments{
\item{fileName}{The name of the file to write to.}
}
\description{
Add the default error report logger
}
\details{
Creates a logger named 'DEFAULT_ERRORREPORT_LOGGER' that writes to a file using the "FATAL" threshold and the
\code{\link{layoutErrorReport}} layout. The file will be overwritten if it is 
older than 60 seconds. The user will be notified that the error report has been
created, and where to find it.
}
