% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{clusterRequire}
\alias{clusterRequire}
\title{Require a package in the cluster}
\usage{
clusterRequire(cluster, package)
}
\arguments{
\item{cluster}{The cluster object.}

\item{package}{The name of the package to load in all nodes.}
}
\description{
Calls the \code{require} function in each node of the cluster.
}
