% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Args.R
\name{convertJsonToSettings}
\alias{convertJsonToSettings}
\title{Converts a JSON string to a settings object}
\usage{
convertJsonToSettings(json)
}
\arguments{
\item{json}{A JSON string.}
}
\value{
An R object as specified by the JSON.
}
\description{
Converts a JSON string to a settings object
}
\details{
Converts a JSON string generated using the \code{\link{convertSettingsToJson}} function to a
settings object, restoring object classes and attributes.
}
