% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Args.R
\name{convertSettingsToJson}
\alias{convertSettingsToJson}
\title{Convert a settings object to a JSON string}
\usage{
convertSettingsToJson(object)
}
\arguments{
\item{object}{R object to be converted.}
}
\value{
A JSON string representing the R object.
}
\description{
Convert a settings object to a JSON string
}
\details{
Convert a settings object to a JSON string, using pretty formatting and preserving object classes
and attributes.
}
