% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Appenders.R
\name{createFileAppender}
\alias{createFileAppender}
\title{Create file appender}
\usage{
createFileAppender(layout = layoutParallel, fileName)
}
\arguments{
\item{layout}{The layout to be used by the appender.}

\item{fileName}{The name of the file to write to.}
}
\description{
Create file appender
}
\details{
Creates an appender that will write to a file.
}
