% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Args.R
\name{excludeFromList}
\alias{excludeFromList}
\title{Exclude variables from a list of objects of the same type}
\usage{
excludeFromList(x, exclude)
}
\arguments{
\item{x}{A list of objects of the same type.}

\item{exclude}{A character vector of names of variables to exclude.}
}
\description{
Exclude variables from a list of objects of the same type
}
