% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{getThreadNumber}
\alias{getThreadNumber}
\title{Return the number of the current thread}
\usage{
getThreadNumber()
}
\value{
Returns the number of the current thread. Returns 0 if this is the main thread.
}
\description{
Return the number of the current thread
}
