% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layouts.R
\name{layoutEmail}
\alias{layoutEmail}
\title{Logging layout for e-mail}
\usage{
layoutEmail(level, message)
}
\arguments{
\item{level}{The level of the message (e.g. "INFO")}

\item{message}{The message to layout.}
}
\description{
A layout function to be used with an e-mail appender. This layout creates a short summary e-mail
message on the event, including stack trace.
}
