% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layouts.R
\name{layoutParallel}
\alias{layoutParallel}
\title{Logging layout for parallel computing}
\usage{
layoutParallel(level, message)
}
\arguments{
\item{level}{The level of the message (e.g. "INFO")}

\item{message}{The message to layout.}
}
\description{
A layout function to be used with an appender. This layout adds the time, thread, level, package
name, and function name to the message.
}
