% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logging.R
\name{layoutStackTrace}
\alias{layoutStackTrace}
\title{Logging layout with stacktrace}
\usage{
layoutStackTrace(level, message)
}
\arguments{
\item{level}{The level of the message (e.g. "INFO")}

\item{message}{The message to layout.}
}
\description{
A layout function to be used with an appender. This layout adds the strack trace to the message.
}
