% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layouts.R
\name{layoutTimestamp}
\alias{layoutTimestamp}
\title{Logging layout with timestamp}
\usage{
layoutTimestamp(level, message)
}
\arguments{
\item{level}{The level of the message (e.g. "INFO")}

\item{message}{The message to layout.}
}
\description{
A layout function to be used with an appender. This layout adds the time to the message.
}
\examples{
appender <- createConsoleAppender(layout = layoutTimestamp)

logger <- createLogger(name = "SIMPLE",
                       threshold = "INFO",
                       appenders = list(appender))
registerLogger(logger)                      
logTrace("This event is below the threshold (INFO)")
logInfo("Hello world")                       
unregisterLogger("SIMPLE")                     
}
