% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Args.R
\name{loadSettingsFromJson}
\alias{loadSettingsFromJson}
\title{Load a settings object from a JSON file}
\usage{
loadSettingsFromJson(fileName)
}
\arguments{
\item{fileName}{Name of the JSON file to load.}
}
\value{
An R object as specified by the JSON.
}
\description{
Load a settings object from a JSON file
}
\details{
Load a settings object from a JSON file, restoring object classes and attributes.
}
