% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logging.R
\name{logError}
\alias{logError}
\title{Log a message at the ERROR level}
\usage{
logError(...)
}
\arguments{
\item{...}{Zero or more objects which can be coerced to character (and which are pasted together
with no separator).}
}
\description{
Log a message at the ERROR level
}
\details{
Log a message at the specified level. The message will be sent to all the registered loggers.
}
