% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParallelLogger.R
\docType{package}
\name{ParallelLogger-package}
\alias{ParallelLogger}
\alias{ParallelLogger-package}
\title{ParallelLogger: Support for Parallel Computation, Logging, and Function Automation}
\description{
Support for parallel computation with progress bar, and option to stop or proceed on errors. Also provides logging to console and disk, and the logging persists in the parallel threads. Additional functions support function call automation with delayed execution (e.g. for executing functions in parallel).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/ParallelLogger/}
  \item \url{https://github.com/OHDSI/ParallelLogger}
  \item Report bugs at \url{https://github.com/OHDSI/ParallelLogger/issues}
}

}
\author{
\strong{Maintainer}: Martijn Schuemie \email{schuemie@ohdsi.org}

Authors:
\itemize{
  \item Marc Suchard
}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
