% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Notification.R
\name{runAndNotify}
\alias{runAndNotify}
\title{Run code and send e-mail notification on error, warning, or completion}
\usage{
runAndNotify(expression, mailSettings, label = "R", stopOnWarning = FALSE)
}
\arguments{
\item{expression}{The expression to run.}

\item{mailSettings}{Arguments to be passed to the send.mail function in the mailR package (except
subject and body).}

\item{label}{A label to be used in the subject to identify a run.}

\item{stopOnWarning}{Stop expression on warning and send notification?}
}
\value{
The output of \code{expression}.
}
\description{
Run code and send e-mail notification on error, warning, or completion
}
\examples{
\dontrun{
mailSettings <- list(from = "someone@gmail.com",
                     to = c("someone_else@gmail.com"),
                     smtp = list(host.name = "smtp.gmail.com",
                                 port = 465,
                                 user.name = "someone@gmail.com",
                                 passwd = "super_secret!",
                                 ssl = TRUE),
                     authenticate = TRUE,
                     send = TRUE)
runAndNotify({
  a <- 1 + 2 + 3
}, mailSettings = mailSettings, label = "My fancy R code")
}

}
