% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Args.R
\name{saveSettingsToJson}
\alias{saveSettingsToJson}
\title{Save a settings object as JSON file}
\usage{
saveSettingsToJson(object, fileName)
}
\arguments{
\item{object}{R object to be saved.}

\item{fileName}{File name where the object should be saved.}
}
\description{
Save a settings object as JSON file
}
\details{
Save a setting object as a JSON file, using pretty formatting and preserving object classes and
attributes.
}
