% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Args.R
\name{selectFromList}
\alias{selectFromList}
\title{Select variables from a list of objects of the same type}
\usage{
selectFromList(x, select)
}
\arguments{
\item{x}{A list of objects of the same type.}

\item{select}{A character vector of names of variables to select.}
}
\description{
Select variables from a list of objects of the same type
}
\examples{

x <- list(
  a = list(name = "John", age = 25, gender = "M"),
  b = list(name = "Mary", age = 24, gender = "F")
)
selectFromList(x, c("name", "age"))

# $a
# $a$name
# [1] "John"
#
# $a$age
# [1] 25
#
#
# $b
# $b$name
# [1] "Mary"
#
# $b$age
# [1] 24
}
