% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loggers.R
\name{addDefaultFileLogger}
\alias{addDefaultFileLogger}
\title{Add the default file logger}
\usage{
addDefaultFileLogger(fileName, name = "DEFAULT_FILE_LOGGER")
}
\arguments{
\item{fileName}{The name of the file to write to.}

\item{name}{A name for the logger.}
}
\description{
Add the default file logger
}
\details{
Creates a logger that writes to a file using the "TRACE" threshold and the
\code{\link{layoutParallel}} layout. The output can be viewed with the built-in log viewer that can
be started using \code{\link{launchLogViewer}}.
}
