% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Appenders.R
\name{createFileAppender}
\alias{createFileAppender}
\title{Create file appender}
\usage{
createFileAppender(
  layout = layoutParallel,
  fileName,
  overwrite = FALSE,
  expirationTime = 60
)
}
\arguments{
\item{layout}{The layout to be used by the appender.}

\item{fileName}{The name of the file to write to.}

\item{overwrite}{Overwrite the file if it is older than the expiration time?}

\item{expirationTime}{Expiration time in seconds}
}
\description{
Create file appender
}
\details{
Creates an appender that will write to a file.
}
