% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{getPhysicalMemory}
\alias{getPhysicalMemory}
\title{Get the total amount of physical memory}
\usage{
getPhysicalMemory()
}
\value{
The number of GB of RAM. Returns NA if the function failed. One GB is 
1,000,000,000 bytes.
}
\description{
Get the total amount of physical memory
}
\examples{
getPhysicalMemory()

}
